CP2K Inputs and Outputs refactored into modules
