+++ transition_sampling/engines/__init__.py
@@ -1,3 +1,3 @@
 from .abstract_engine import AbstractEngine, ShootingResult
-from .CP2K_engine import CP2KEngine
+from .cp2k.CP2K_engine import CP2KEngine
 from .plumed import PlumedInputHandler
