+++ transition_sampling/tests/engine_tests/cp2k_tests/test_CP2K_engine.py
@@ -1,5 +1,4 @@
 import copy
-import filecmp
 import os
 import shutil
 import tempfile
@@ -8,13 +7,11 @@ from unittest import TestCase
 import numpy as np
 
 from engines import CP2KEngine
-from engines.CP2K_engine import write_cp2k_input, CP2KOutputHandler
 
 ENG_STR = "cp2k"
 CUR_DIR = os.path.dirname(__file__)
-TEST_INPUT = os.path.join(CUR_DIR, "test_data/cp2k/test_cp2k.inp")
-TEST_OUTPUT = os.path.join(CUR_DIR, "test_data/cp2k/test_cp2k_warnings.out")
-TEST_PLUMED_FILE = os.path.join(CUR_DIR, "test_data/cp2k/test_plumed.dat")
+TEST_INPUT = os.path.join(CUR_DIR, "test_data/test_cp2k.inp")
+TEST_PLUMED_FILE = os.path.join(CUR_DIR, "test_data/test_plumed.dat")
 TEST_CMD = "test cmd"
 
 CORRECT_INPUTS = {"engine": ENG_STR,
@@ -103,6 +100,9 @@ class TestCP2KEngineValidation(TestCase):
                              msg="Test input should be valid")
 
 
+# TODO: Most of these tests are for the interface. When more engines are added,
+# we should have one standard system (ex 2 Ar atoms) and run all engines with
+# these tests
 class TestCP2KEngineAtoms(CP2KEngineTestCase):
     def test_atoms_getting(self):
         """
@@ -144,50 +144,13 @@ class TestCP2KEnginePositions(CP2KEngineTestCase):
 
     def test_set_positions_valid(self):
         """
-        Assign valid positions and check the internal representation of them
+        Assign valid positions to ensure it works. Specifics are tested by the
+        CP2KInputHandler
         """
         pos = np.array([[1.0021, 123.123, 6.23123],
                         [8.12, 6.12381, 0.1232]])
 
         self.engine.set_positions(pos)
-
-        self._compare_positions(pos, self.engine)
-
-    def test_set_positions_and_write(self):
-        """
-        Assign positions, write to a file, load into a new engine, and see if
-        they match
-        """
-        pos = np.array([[1.0021, 123.123, 6.23123],
-                        [8.12, 6.12381, 0.1232]])
-
-        self.engine.set_positions(pos)
-        with tempfile.NamedTemporaryFile() as temp_file:
-            write_cp2k_input(self.engine.cp2k_inputs, temp_file.name)
-
-            # Load positions from the saved temp file
-            copied_inputs = copy.deepcopy(CORRECT_INPUTS)
-            copied_inputs["cp2k_inputs"] = temp_file.name
-            new_engine = CP2KEngine(copied_inputs)
-
-            self._compare_positions(pos, new_engine)
-
-    def _compare_positions(self, expected, engine):
-        """
-        Compare expected positions to those actually stored by an engine
-        :param expected: Array of expected positions
-        :param engine: Engine to compare to
-        """
-        actual = \
-            engine.cp2k_inputs["+force_eval"][0]["+subsys"]["+coord"]["*"]
-
-        # Iterate over each stored and assigned position to compare
-        for s, p in zip(actual, expected):
-            # Convert string representation to list of floats
-            split_list = [float(num) for num in s[3:].split()]
-            # Convert numpy array to list for assertListEquals
-            p = p.tolist()
-            self.assertListEqual(p, split_list, "Positions were not equal")
 
 
 class TestCP2KEngineVelocities(CP2KEngineTestCase):
@@ -214,96 +177,8 @@ class TestCP2KEngineVelocities(CP2KEngineTestCase):
 
     def test_set_velocities_valid(self):
         """
-        Assign valid velocities and check the internal representation of them
+        Assign valid velocities to ensure it works. Specifics checked by
+        CP2KInputsHandler
         """
         vel = np.array([[1.0021, 123.123, 6.23123],
                         [8.12, 6.12381, 0.1232]])
-
-        self.engine.set_velocities(vel)
-        self._compare_velocities(vel, self.engine)
-
-    def test_set_velocities_and_write(self):
-        """
-        Assign velocities, write to a file, load into a new engine, and see if
-        they match
-        """
-        vel = np.array([[1.0021, 123.123, 6.23123],
-                        [8.12, 6.12381, 0.1232]])
-
-        self.engine.set_velocities(vel)
-
-        with tempfile.NamedTemporaryFile() as temp_file:
-            write_cp2k_input(self.engine.cp2k_inputs, temp_file.name)
-
-            # Load velocities from the saved temp file
-            copied_inputs = copy.deepcopy(CORRECT_INPUTS)
-            copied_inputs["cp2k_inputs"] = temp_file.name
-            new_engine = CP2KEngine(copied_inputs)
-
-            self._compare_velocities(vel, new_engine)
-
-    def test_flip_velocities(self):
-        """Test that flipping velocities works"""
-        vel = np.array([[1.0021, 123.123, 6.23123],
-                        [8.12, 6.12381, 0.1232]])
-
-        self.engine.set_velocities(vel)
-        self.engine.flip_velocity()
-
-        self._compare_velocities(-1 * vel, self.engine)
-
-    def _compare_velocities(self, expected, engine):
-        """
-        Compare the expected values of a velocity to those actually stored by
-        an engine
-        :param expected: array of expected velocities
-        :param engine: engine to check if velocities match
-        """
-        # Internal Representation of stored positions for CP2K
-        actual = \
-            engine.cp2k_inputs["+force_eval"][0]["+subsys"]["+velocity"]["*"]
-
-        # Iterate over each stored and assigned position to compare
-        for s, v in zip(actual, expected):
-            # Convert numpy array to list for assertListEquals
-            v = v.tolist()
-            self.assertListEqual(v, s, "Velocities were not equal")
-
-
-class TestCP2KEngineWritePlumed(CP2KEngineTestCase):
-    """Tests for CP2KEngine interactions with plumed"""
-    def test_plumed_set_file_in_cp2k_inputs(self):
-        """Ensure the plumed file name gets set in the inputs"""
-        file_name = "TEST_FILE"
-        self.engine.set_plumed_file(file_name)
-
-        # access the internal representation
-        metad = self.engine.cp2k_inputs["+motion"]["+free_energy"]["+metadyn"]
-
-        self.assertEqual(metad["plumed_input_file"], file_name,
-                         msg="Plumed file name not correct")
-
-        self.assertTrue(metad["use_plumed"], msg="Use plumed was not true")
-
-
-class TestCP2KOutputHandler(TestCase):
-
-    def setUp(self) -> None:
-        test_dir = os.path.join(CUR_DIR, "test_data/cp2k")
-        self.out_handler = CP2KOutputHandler("test_cp2k_warnings", test_dir)
-
-    def test_output_handler_builds_path(self):
-        """Test that output file name gets built correctly"""
-        self.assertEqual(self.out_handler.get_out_file(),
-                         TEST_OUTPUT)
-
-    def test_output_handler_copies_file(self):
-        with tempfile.NamedTemporaryFile() as temp_file:
-            self.out_handler.copy_out_file(temp_file.name)
-
-            self.assertTrue(filecmp.cmp(self.out_handler.get_out_file(),
-                                        temp_file.name), "files were not equal")
-
-    def test_output_handler_catches_warnings(self):
-        self.assertEqual(len(self.out_handler.check_warnings()), 1,
-                         "Warnings were not caught")
